% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.R
\name{annotation_compass}
\alias{annotation_compass}
\title{Adding Text to a ggplot2 Object}
\usage{
annotation_compass(
  label,
  position = c("N", "NE", "E", "SE", "S", "SW", "W", "NW"),
  padding = grid::unit(c(0.5, 0.5), "line"),
  ...
)
}
\arguments{
\item{label}{character. Label of the legend}

\item{position}{Position of the legend. Options are "N", "NE", "E", "SE", "S", "SW", "W", "NW".}

\item{padding}{A numeric vector or unit object specifying x-values and y-values. Default is grid::unit(c(0.5, 0.5), "line")}

\item{...}{Other arguments}
}
\value{
A ggplot2 object with legend.
}
\description{
This function adds text to the specified places (using compass style) of a ggplot2 object.
}
\examples{
\dontrun{
g + annotation_compass(label = "sometext", position = "NE") ## Do not run.

## To do the same, you can always use:
g + geom_text(data = data.frame(), aes(label = "sometext",
              x = -Inf, y = Inf), hjust = 0, vjust = 1) ## Do not run.
g + ggplot2::annotate(geom = "text", label = "sometext",
                      x = Inf, y = Inf, hjust = 1,
                      vjust = 1, size = 24) ## Do not run.
}

}
\references{
For more information see \href{https://stackoverflow.com/questions/47916307/specify-position-of-geom-text-by-keywords-like-top-bottom-left-right}{here}.
}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
