% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{Ask.User.YN.Question}
\alias{Ask.User.YN.Question}
\title{Interactive Yes/No Questions}
\usage{
Ask.User.YN.Question(question, GUI = TRUE, add.lines.before = TRUE)
}
\arguments{
\item{question}{string. Your question as a character string.}

\item{GUI}{logical. TRUE for using GUI or FALSE for using console.}

\item{add.lines.before}{logical. Simply add some line after your questions.}
}
\value{
logical.
}
\description{
This function creates an interactive yes/no question.
}
\examples{
\dontrun{
if (Ask.User.YN.Question("Your question goes here?", GUI = FALSE, add.lines.before = TRUE)) {
    Some Code Here
} else {
    More Code Here
}
}

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
