% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{convert.data}
\alias{convert.data}
\alias{Dash.Convert.NA}
\alias{Dash.Convert.Zero}
\alias{Zero.Convert.NA}
\alias{Null.Convert.NA}
\alias{NA.Convert.Zero}
\alias{Negative.Convert.NA}
\alias{NaN.Convert.Zero}
\alias{NaN.Convert.NA}
\title{Functions for Converting Data}
\usage{
Dash.Convert.NA(data, variable)

Dash.Convert.Zero(data, variable)

Zero.Convert.NA(data, variable)

Null.Convert.NA(data, variable)

NA.Convert.Zero(data, variable)

Negative.Convert.NA(data, variable)

NaN.Convert.Zero(data, variable)

NaN.Convert.NA(data, variable)
}
\arguments{
\item{data}{data.frame. Data as in data.frame class.}

\item{variable}{character. A single string which is a column name in the data.}
}
\value{
A data frame.
}
\description{
This set of functions converts given data by the selected column with some specific purposes. The details are below.
}
\details{
All conversions are in character. See below.
\itemize{
\item Dash.Convert.NA: Converts "-" into NA.
\item Dash.Convert.Zero: Converts "-" into "0".
\item Zero.Convert.NA: Converts "0" into NA.
\item Null.Convert.NA: Converts "null" into NA.
\item NA.Convert.Zero: Converts NA into "0".
\item Negative.Convert.NA: Converts negative value into NA.
\item NaN.Convert.Zero: Converts NaN into "0".
\item NaN.Convert.NA: Converts NaN into NA.
}
}
\examples{
\dontrun{
Dash.Convert.NA(data, "variable.name") ## Do not run.
Dash.Convert.Zero(data, "variable.name") ## Do not run.
Zero.Convert.NA(data, "variable.name") ## Do not run.
Null.Convert.NA(data, "variable.name") ## Do not run.
NA.Convert.Zero(data, "variable.name") ## Do not run.
Negative.Convert.NA(data, "variable.name") ## Do not run.
NaN.Convert.Zero(data, "variable.name") ## Do not run.
NaN.Convert.NA(data, "variable.name") ## Do not run.
}

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Omer Kara}
}
