% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{convert.degree.radian}
\alias{convert.degree.radian}
\alias{deg2rad}
\alias{rad2deg}
\title{Functions for Converting Degrees to/from Radians}
\usage{
deg2rad(deg)

rad2deg(rad)
}
\arguments{
\item{deg}{numeric. A numeric value or vector of degrees.}

\item{rad}{numeric. A numeric value or vector of radians.}
}
\value{
Numeric value or vector.
}
\description{
This set of functions either converts the given value from degrees-to-radians or radians-to-degrees.
}
\details{
\itemize{
\item deg2rad: Converts degrees to radians.
\item rad2deg: Converts radians to degrees.
}
}
\note{

}
\examples{
deg2rad(100)
rad2deg(100)

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
