% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{convert.distance.degree}
\alias{convert.distance.degree}
\alias{km2d}
\alias{ml2d}
\alias{d2km}
\alias{d2ml}
\title{Functions for Converting Distances to/from Degrees}
\usage{
km2d(km, base.latitude = 38.280479)

ml2d(ml, base.latitude = 38.280479)

d2km(d, base.latitude = 38.280479)

d2ml(d, base.latitude = 38.280479)
}
\arguments{
\item{km}{numeric. A single numeric km distance.}

\item{base.latitude}{numeric. A single numeric base latitude. Base latitude is generally selected as the mean latitude of your shapefile.}

\item{ml}{numeric. A single numeric mile distance.}

\item{d}{numeric. A single numeric degrees.}
}
\value{
Numeric value.
}
\description{
This set of functions converts given distances either from km-to-degrees and miles-to-degrees or degrees-to-km and degrees-to-miles.
}
\details{
\itemize{
\item km2d: Converts km to degrees by the selected latitude degree.
\item ml2d: Converts miles to degrees by the selected latitude degree.
\item d2km: Converts degrees to kilometers by the selected latitude degree.
\item d2ml: Converts degrees to miles by the selected latitude degree.
}
}
\examples{
km2d(100)
ml2d(100)
d2km(100)
d2ml(100)

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
