% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{convert.distance}
\alias{convert.distance}
\alias{km2ml}
\alias{ml2km}
\title{Functions for Converting Distances}
\usage{
km2ml(km)

ml2km(ml)
}
\arguments{
\item{km}{numeric. A numeric value or vector of km distance.}

\item{ml}{numeric. A numeric value or vector of mile distance.}
}
\value{
Numeric value or vector.
}
\description{
This set of functions converts given distances either from km-to-miles or miles-to-km.
}
\details{
\itemize{
\item km2ml: Converts kilometers to miles.
\item ml2km: Converts miles to kilometers.
}
}
\examples{
km2ml(1.6)
ml2km(1)

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
