% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{dataStr}
\alias{dataStr}
\title{Structure of Datasets}
\usage{
dataStr(package = NULL, df = FALSE, ...)
}
\arguments{
\item{package}{character. Package name. Default: NULL}

\item{df}{logical. Give information only about all data.frame objects? DEFAULT: FALSE}

\item{...}{Arguments passed from other functions.}
}
\value{
Invisible data.frame. Mainly prints via \code{\link{message}} in a for loop.
}
\description{
This function lists the data sets (with their structure) in specified R packages.
}
\examples{
\dontrun{
}

}
\references{
This function is taken from \href{https://github.com/brry/berryFunctions/blob/master/R/dataStr.R}{here}.
}
\seealso{
\code{\link{str}}
}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
