% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{Decimal.Num.Count}
\alias{Decimal.Num.Count}
\title{Number of Decimal Places}
\usage{
Decimal.Num.Count(x)
}
\arguments{
\item{x}{Number in character or in numeric class.}
}
\value{
An integer.
}
\description{
This function gives the number of decimal places of a number written as character string or numeric object.
}
\note{
Note that when object given is a character then trailing zeros are not dropped and all the decimal places are returned. For example, Decimal.Num.Count("0.10") gives 2 decimal places instead of 2.
}
\examples{
Decimal.Num.Count(0.1)
Decimal.Num.Count("0.1")
Decimal.Num.Count("0.10")
Decimal.Num.Count(0.10)

Decimal.Num.Count(0.12345)
Decimal.Num.Count("0.12345")

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
