% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{diffPad}
\alias{diffPad}
\title{Difference Function with Padding}
\usage{
diffPad(x)
}
\arguments{
\item{x}{ts object or numeric vector.}
}
\value{
A first difference taken ts object or numeric vector with padding.
}
\description{
This function takes the first difference of a given time series data or vector and pads it with the selected lag length while maintaining the attributes of the input data. ts object or vector can be used.
}
\details{

}
\note{

}
\examples{
values <- c(1:10)
x <- values
x.ts <- ts(data.frame(V1 = values))
diffPad(x)
diffPad(x.ts)

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
