% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{DigitsByRows}
\alias{DigitsByRows}
\title{Significant Digits By Row}
\usage{
DigitsByRows(df, digits)
}
\arguments{
\item{df}{A data frame.}

\item{digits}{The requested digits in order of rows.}
}
\value{
A data frame.
}
\description{
This function prints significant digits by row in a data frame.
}
\examples{
\dontrun{
DigitsByRows(df, digits = c(0, 2)) ## Do not run.
}

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
