% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.R
\name{g_legend}
\alias{g_legend}
\title{Extracting Legend of a ggplot2 Object}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{A ggplot2 object.}
}
\value{
A ggplot2 legend.
}
\description{
This function extracts legend from ggplot2 plots.
}
\details{

}
\note{

}
\examples{
\dontrun{
g <- ggplot(...) ## Do not run.
g_legend() ## Do not run.
}

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
