% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{Head.Tail}
\alias{Head.Tail}
\title{Head and Tail of a Data Frame}
\usage{
Head.Tail(x, Select)
}
\arguments{
\item{x}{data.frame. Data as in data.frame class.}

\item{Select}{integer. Number of the first and the last items in the data frame.}
}
\value{
A matrix look alike object.
}
\description{
This function shows the first and last selected items of a data frame, then presents these together.
}
\examples{
\dontrun{
Head.Tail(data, 5) ## Do not run.
Head.Tail(data, 10) ## Do not run.
}

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
