% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{hoCoef}
\alias{hoCoef}
\title{t-Tests for Class lm Objects}
\usage{
hoCoef(object, term = 2, bo = 0, alt = c("two.sided", "less", "greater"))
}
\arguments{
\item{object}{Model.}

\item{term}{Order of the parameter in the model.}

\item{bo}{Alternative hypothesis value.}

\item{alt}{Alternative hypothesis type.}
}
\value{
A t-test result table.
}
\description{
This function performs t-Tests for class lm objects .
}
\examples{
\dontrun{
hoCoef(your.model, term = 2, bo = 0, alt = "two.sided") ## Do not run.
hoCoef(your.model, term = 2, bo = -1, alt = "less") ## Do not run.
hoCoef(your.model, term = 4, bo = 3, alt = "greater") ## Do not run.
}

}
\references{
The hoCoef function is taken directly from the FSA package (version 0.8.32). For more information see \href{https://cran.r-project.org/web/packages/FSA/index.html}{here}.
}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Omer Kara}
}
