% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.R
\name{human.numbers}
\alias{human.numbers}
\alias{human_numbers}
\alias{human_num}
\alias{human_per}
\alias{human_gbp}
\alias{human_usd}
\alias{human_euro}
\alias{human_tl}
\title{Human Numbers For ggplot2 Graph Axis}
\usage{
human_numbers(x = NULL, smbl = "", signif = 1)

human_num(x)

human_per(x)

human_gbp(x)

human_usd(x)

human_euro(x)

human_tl(x)
}
\arguments{
\item{x}{numeric. a numeric vector to format.}

\item{smbl}{character. a symbol you'd like to prefix your numbers by e.g. "$".}

\item{signif}{numeric. the number of significant places you want the function to return.}
}
\value{
A character vector the same length as the input vector
}
\description{
This set of functions formats the numbers in a ggplot2 graph axis so they are easily readable for humans. Use this function in a ggplot2 object for labels where you might use the comma or percent functions from the Scales package. Function checks whether numbers are positive or negative. It allows up to 1 significant figure and sapply used for element-wise application of the humanity function as a vector may include numbers where billions, millions or thousands are appropriate.
}
\details{
All conversions are in character. See below.
\itemize{
\item human_numbers: Main function with adjustable symbol and significant places.
\item human_num: For no symbol.
\item human_per: For percentage symbol.
\item human_gbp: For Pound symbol.
\item human_usd: For Dollar symbol.
\item human_euro: For Euro symbol.
\item human_tl: For TL symbol.
}
}
\examples{
human_numbers(c(1000000 , 1500000, 10000000000))
human_numbers(c(1.200000e+05, -2.154660e+05, 2.387790e+05, 4.343500e+04 ,5.648675e+12), "$")

\dontrun{
ggplot2 + scale_y_continuous(labels = human_num)
ggplot2 + scale_x_continuous(labels = human_gbp)
ggplot2 + scale_x_continuous(labels = human_tl)
}

}
\references{
These functions are taken from \href{https://github.com/fdryan/R/blob/master/ggplot2_formatter.r}{here}.
}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
