% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{Impute.NA.Mean}
\alias{Impute.NA.Mean}
\title{Impute Missing Values with Mean}
\usage{
Impute.NA.Mean(x)
}
\arguments{
\item{x}{numeric. a numeric vector.}
}
\value{
A numeric vector.
}
\description{
This function fills NA values with mean value.
}
\details{

}
\note{

}
\examples{
Impute.NA.Mean(c(1:5, rep(NA, 1)))

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
