% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{lagPad}
\alias{lagPad}
\title{Lag Function with Padding}
\usage{
lagPad(x, k)
}
\arguments{
\item{x}{numeric. A numeric vector.}

\item{k}{numeric. A numeric lag value.}
}
\value{
A lag taken numeric vector with padding.
}
\description{
This function pads the lag taken numeric vector with NA values.
}
\details{

}
\note{

}
\examples{
x <- 1:10
x.t1 <- lagPad(x, 2)
cbind(x, x.t1)

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
