% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{Load.Install}
\alias{Load.Install}
\title{Load or Install/Load Packages}
\usage{
Load.Install(Package.Names, Quiet = FALSE, Update.All = FALSE)
}
\arguments{
\item{Package.Names}{A vector of strings. The package names to be installed/loaded.}

\item{Quiet}{logical. If TRUE, suppress the output.}

\item{Update.All}{logical. If TRUE, updates the specified packages.}
}
\value{
Install and/or loads the specified packages.
}
\description{
This function either loads or installs/loads the specified packages.
}
\details{
Packages from GitHub can also be installed/loaded.
}
\note{

}
\examples{
\dontrun{
Load.Install(Package.Names = "plyr")
Load.Install(Package.Names = c("plyr", "dplyr"))
Load.Install(c("plyr", "dplyr", "tidyr", "stringr", "stringi", "Hmisc"))
Load.Install(c("plyr", "dplyr"), Quiet = TRUE, Update.All = TRUE)

Load.Install("FinYang/tsdl")
}

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
