% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{Load.Packages}
\alias{Load.Packages}
\title{Load Packages}
\usage{
Load.Packages(Package.Names, Quiet = FALSE)
}
\arguments{
\item{Package.Names}{A vector of strings. The package names to be loaded.}

\item{Quiet}{logical. If TRUE, suppress the output.}
}
\value{
Loads the specified packages.
}
\description{
This function loads the specified packages.
}
\details{
Packages from GitHub can also be loaded.
}
\note{

}
\examples{
\dontrun{
Load.Packages(Package.Names = "plyr")
Load.Packages(Package.Names = c("plyr", "dplyr"))
Load.Packages(c("plyr", "dplyr", "tidyr", "stringr", "stringi", "Hmisc"))
Load.Packages(c("plyr", "dplyr"), Quiet = TRUE)

Load.Packages("tsdl") ## Github package.
}

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
