% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{logPercent}
\alias{logPercent}
\title{Percentage Change for Log Variables}
\usage{
logPercent(Variable, Percent)
}
\arguments{
\item{Variable}{numeric. A numeric value or vector in logarithmic form.}

\item{Percent}{numeric. A numeric value representing percentage increase in the level form.}
}
\value{
A value in the logarithmic form which makes sure that the level form value has increased by the specified percentage.
}
\description{
This function makes sure that the level variable of interest increases/decreases, 10\% i.e., while you are dealing with the log form.
}
\examples{
x <- 100
percent <- 10
log(x)
logPercent(log(x), percent)
exp(logPercent(log(x), percent))

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
