% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal_functions.R
\name{Min.Sell.Buy.Price}
\alias{Min.Sell.Buy.Price}
\title{Minimum Sell or Buy Prices}
\usage{
Min.Sell.Buy.Price(Price, Fee.Per = 0.1)
}
\arguments{
\item{Price}{numeric. Buy or sell price.}

\item{Fee.Per}{numeric. Fee percentage.}
}
\value{
Three data.frames called "Time.Detail", "Total.Days", and "Total.Weeks".
}
\description{
This function gives you the minimum sell price after you buy or the minimum buy price after you sell with a selected fee percentage which makes you break even after the fee.
}
\details{
This is a personal function of the package author. So, it won't make sense to others.
}
\note{

}
\examples{
Min.Sell.Buy.Price(Price = 60, Fee.Per = 0.1)
Min.Sell.Buy.Price(Price = 60)
Min.Sell.Buy.Price(Price = 60, Fee.Per = 0.08)

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
