% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okara.R
\docType{package}
\name{okara}
\alias{okara}
\alias{okara-package}
\title{Omer Kara's Personal Utility Functions}
\description{
The okara package consists of Omer Kara's frequently used utility functions about the following topics. Data manipulation, data utility, graphic/plotting, estimation, spatial analysis, package installing/loading and some other utility functions.
}
\author{
Omer Kara \email{omer.kara.ylsy@gmail.com}
}
