% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{Percent.Rank}
\alias{Percent.Rank}
\title{Percentage Rank}
\usage{
Percent.Rank(data, signif = 3, percent = TRUE)
}
\arguments{
\item{data}{numeric vector. Vector of numeric values.}

\item{signif}{integer. A numeric value representing significant decimal digits.}

\item{percent}{logical. Default is TRUE. If FALSE, the results are in ratio. If TRUE, the results are in percentage.}
}
\value{
An numeric vector.
}
\description{
This function calculates percentage rank same as the "PERCENTRANK.INC()" in Excel.
}
\examples{
\dontrun{
Percent.Rank(data, signif = 3, percent = TRUE)
Percent.Rank(data, signif = 5, percent = FALSE)
}

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
