% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{Proceed.or.Stop}
\alias{Proceed.or.Stop}
\title{Tells to Proceed or Stop}
\usage{
Proceed.or.Stop(result)
}
\arguments{
\item{result}{A result with a result of TRUE or FALSE.}
}
\value{
A text message.
}
\description{
This function produces warning message if there is anything wrong with your code.
}
\examples{
Proceed.or.Stop(5 < 6)
Proceed.or.Stop(5 < 3)
Proceed.or.Stop(5 == 5)

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
