% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{roundN}
\alias{roundN}
\title{Function for Rounding with \code{\link[exams]{round2}}}
\usage{
roundN(x, digits)
}
\arguments{
\item{x}{numeric. A numeric vector.}

\item{digits}{numeric. Rounding digits.}
}
\value{
Numeric value or vector.
}
\description{
This function simplifies and specializes the \code{\link[exams]{round2}} function in exam package. Function rounds the given number to the decimal points specified in argument digits.
}
\details{
The function \code{\link[exams]{round2}} does what is known in German as kaufmaennisches Runden (rounding away from zero for trailing 5s).
}
\examples{
roundN(2.67896, digits = 0)
roundN(2.67896, digits = 1)
roundN(2.67896, digits = 2)
roundN(2.67896, digits = 3)
roundN(2.67896, digits = 4)

}
\references{
For more information see \code{\link[exams]{round2}}.
}
\seealso{
\code{\link[exams]{round2}}
}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Omer Kara}
}
