% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{roundUp}
\alias{roundUp}
\title{Function for Rounding Up with a Given Sensitivity}
\usage{
roundUp(x, to = 1)
}
\arguments{
\item{x}{numeric. A numeric vector.}

\item{to}{numeric. Sensitivity of the rounding.}
}
\value{
Numeric value or vector.
}
\description{
Function rounds the numbers up with the given sensitivity.
}
\examples{
roundUp(2.67896, to = 1)
roundUp(2.67896, to = 0.1)
roundUp(2.67896, to = 0.01)
roundUp(2.67896, to = 0.001)

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
