% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{roundX}
\alias{roundX}
\alias{round0}
\alias{round1}
\alias{round2}
\alias{round3}
\alias{round4}
\title{Functions for Rounding with \code{\link[exams]{round2}}}
\usage{
round0(x)

round1(x)

round2(x)

round3(x)

round4(x)
}
\arguments{
\item{x}{numeric. A numeric vector.}
}
\value{
Numeric value or vector.
}
\description{
This set of functions simplify and specialize the \code{\link[exams]{round2}} function in exam package. Each function rounds the given number to the decimal points specified in the function itself.
}
\details{
The function \code{\link[exams]{round2}} does what is known in German as kaufmaennisches Runden (rounding away from zero for trailing 5s).
}
\examples{
round0(2.67896)
round1(2.67896)
round2(2.67896)
round3(2.67896)
round4(2.67896)

}
\references{
For more information see \code{\link[exams]{round2}}.
}
\seealso{
\code{\link[exams]{round2}}
}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
