% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{sampled}
\alias{sampled}
\title{Bug free Redefinition of Sample Function}
\usage{
sampled(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose, or a positive integer.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled.}
}
\value{
For sample a vector of length size with elements drawn from either x or from the integers 1:x.
}
\description{
If only one value is given in \code{\link[base]{sample}} function, it chooses a value between 1 and the number you specified. The redefined sample function solves this problem.
}
\examples{
sampled(2, 1)
sampled(1:5, 2)

}
\references{
For more information see \code{\link[base]{sample}}.
}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
