% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{SignPrint}
\alias{SignPrint}
\title{Prints The Sign of a Value}
\usage{
SignPrint(x)
}
\arguments{
\item{x}{numeric. A single numeric value.}
}
\value{
Character string either "+" or "-".
}
\description{
This function shows the sign of a given single numeric value.
}
\details{

}
\note{

}
\examples{
SignPrint(2)
SignPrint(-2)

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
