% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_functions.R
\name{SummaryR.lm}
\alias{SummaryR.lm}
\title{Summary of OLS with Robust Std Errors}
\usage{
SummaryR.lm(model, type = c("hc3", "hc0", "hc1", "hc2", "hc4"), ...)
}
\arguments{
\item{model}{Name of the model estimated with \code{\link[stats]{lm}} function.}

\item{type}{Type of the robust std error calculation. See \code{\link[car]{hccm}} function in car package.}

\item{...}{Arguments passed from other functions.}
}
\value{
A linear model summary.
}
\description{
This function gives summary for OLS with different type of robust std errors.
}
\examples{
\dontrun{
SummaryR.lm(my.model, type = "hc4") ## Do not run.
SummaryR.lm(your.model, type = "hc1") ## Do not run.
}

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Omer Kara}
}
