% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal_functions.R
\name{Time.Keeping}
\alias{Time.Keeping}
\title{Keeping Time for Hourly Jobs}
\usage{
Time.Keeping(
  Hours.Punch,
  First.Date = NULL,
  Last.Date = NULL,
  Exclude.Weekends = FALSE,
  Target.Hours = NULL
)
}
\arguments{
\item{Hours.Punch}{character. The path of the time keeping .txt file.}

\item{First.Date}{character. Default if NULL. The first date of time keeping. If NULL, then it will be generated from the data.}

\item{Last.Date}{character. Default if NULL. The last date of time keeping. If NULL, then it will be generated from the data.}

\item{Exclude.Weekends}{logical. If TRUE, the weekends will be excluded from the data.}

\item{Target.Hours}{numeric. Default is NULL. If a numeric value is given, the remaining hours for the current week is messaged.}
}
\value{
Three data.frames called "Time.Detail", "Total.Days", "Total.Weeks", and a vector "total.hours".
}
\description{
This function keeps the punched hours for hourly jobs.
}
\details{
This is a personal function of the package author. So, it won't make sense to others.
}
\examples{
\dontrun{
Time.Keeping(Hours.Punch = "./hours_punch.txt")
Time.Keeping(Hours.Punch = "./hours_punch.txt", Exclude.Weekends = FALSE, Target.Hours = NULL)
Time.Keeping(Hours.Punch = "./hours_punch.txt", First.Date = "2016-08-16", Target.Hours = 20)
}

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Omer Kara}
}
