% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{Cubic.Spline.Interpolation}
\alias{Cubic.Spline.Interpolation}
\title{Cubic Spline Interpolation}
\usage{
Cubic.Spline.Interpolation(Data, Column.Names)
}
\arguments{
\item{Data}{data.frame. Data as in data.frame class.}

\item{Column.Names}{character. Name of the columns to be interpolated.}
}
\value{
An data.frame called "data.interpolate".
}
\description{
This function interpolates the selected columns of a data frame with cubic spline interpolation.
}
\details{
The aim of this function is to interpolate the values between two entries. The argument maxgap in \code{\link[zoo]{na.spline}} function is internally selected in this function to interpolate all missing values. \code{na.rm} argument in \code{\link[zoo]{na.spline}} function is selected as FALSE since other NA values should be kept, if there is any.
}
\examples{
\dontrun{
Cubic.Spline.Interpolation(data, "Open")
data.interpolate

Cubic.Spline.Interpolation(data, c("Open", "High", "Low", "Close"))
data.interpolate
}

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Omer Kara}
}
