% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal_functions.R
\name{Min.Sell.Buy.Price}
\alias{Min.Sell.Buy.Price}
\title{Minimum Sell or Buy Prices}
\usage{
Min.Sell.Buy.Price(Price, Fee.Per = 0.1, Round = 2)
}
\arguments{
\item{Price}{numeric. Buy or sell price.}

\item{Fee.Per}{numeric. Fee percentage. Default is 0.1.}

\item{Round}{numeric. Decimal rounding. Default is 2.}
}
\value{
Two messages: one for BUY Price, one for SELL Price.
}
\description{
This function gives you the minimum sell price after you buy or the minimum buy price after you sell with a selected fee percentage which makes you break even after the fee.
}
\details{
This is a personal function of the package author. So, it won't make sense to others.
}
\examples{
Min.Sell.Buy.Price(Price = 60)
Min.Sell.Buy.Price(Price = 60, Fee.Per = 0.08)
Min.Sell.Buy.Price(Price = 60, Fee.Per = 0.1, Round = 3)

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
