% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{roundDown}
\alias{roundDown}
\title{Function for Rounding Down with a Given Sensitivity}
\usage{
roundDown(x, to = 1)
}
\arguments{
\item{x}{numeric. A numeric vector.}

\item{to}{numeric. Sensitivity of the rounding.}
}
\value{
Numeric value or vector.
}
\description{
Function rounds the numbers Down with the given sensitivity.
}
\details{

}
\note{

}
\examples{
roundDown(2.67896, to = 1)
roundDown(2.67896, to = 0.1)
roundDown(2.67896, to = 0.01)
roundDown(2.67896, to = 0.001)

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
