% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{truncN}
\alias{truncN}
\alias{trunc0}
\alias{trunc1}
\alias{trunc2}
\alias{trunc3}
\alias{trunc4}
\title{Functions for Truncating without Round}
\usage{
trunc0(x)

trunc1(x)

trunc2(x)

trunc3(x)

trunc4(x)
}
\arguments{
\item{x}{numeric. A numeric vector.}
}
\value{
Numeric value or vector.
}
\description{
This set of functions truncates the numbers without rounding. Each function truncates the given number to the decimal points specified in the function itself.
}
\details{

}
\note{

}
\examples{
trunc0(2.67896)
trunc1(2.67896)
trunc2(2.67896)
trunc3(2.67896)
trunc4(2.67896)

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
