% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{truncN}
\alias{truncN}
\title{Function for Truncating without Round}
\usage{
truncN(x, digits)
}
\arguments{
\item{x}{numeric. A numeric vector.}

\item{digits}{numeric. Truncating digits.}
}
\value{
Numeric value or vector.
}
\description{
This function truncates the numbers without rounding. Function truncates the given number to the decimal points specified in argument digits. Note that rounding is performed with 1 extra digits to correctly truncate the values as much as possible.
}
\details{

}
\note{

}
\examples{
truncN(2.67896, digits = 0)
truncN(2.67896, digits = 1)
truncN(2.67896, digits = 2)
truncN(2.67896, digits = 3)
truncN(2.67896, digits = 4)

}
\references{

}
\seealso{

}
\author{
\href{mailto:omer.kara.ylsy@gmail.com}{Ömer Kara}
}
